requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
anndata==0.7.8
anyio==3.5.0
apptools==5.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
astor==0.8.1
astroid==2.9.3
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bash==0.6
bcrypt==3.2.0
binaryornot==0.4.4
black==22.1.0
bleach==4.1.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
common==0.1.2
configobj==5.0.6
contextlib2==21.6.0
cookiecutter==1.7.3
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20200713
dm-env==1.5
dm-tree==0.1.7
docutils==0.17.1
dual==0.0.10
entrypoints==0.4
envisage==6.0.1
fastdist==1.1.3
filelock==3.6.0
flake8==4.0.1
flatbuffers==2.0
frozenlist==1.3.0
future==0.18.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gym==0.23.0
gym-notices==0.0.5
h5py==3.6.0
hickory==1.0.0
hyperopt==0.1.2
idna==3.3
imagecodecs==2021.11.20
imageio==2.16.1
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.9.1
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
json-tricks==3.15.5
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter-packaging==0.11.1
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
learning3d==0.1.2
libclang==13.0.0
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.3.3
matplotlib-inline==0.1.3
mayavi==4.7.4
mccabe==0.6.1
memory-profiler==0.60.0
mistune==0.8.4
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
natsort==8.1.0
nb-conda==0.1.0
nb-conda-kernels==2.2.2
nbclassic==0.3.7
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
nibabel==3.2.2
ninja==1.10.2.3
nni==2.6.1
notebook==5.2.2
notebook-shim==0.1.0
numba==0.55.1
numexpr==2.8.1
numpy==1.21.5
numpydoc==1.2
oauthlib==3.2.0
open3d==0.15.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
orca==1.6
packaging==21.3
pandas==1.2.2
pandocfilters==1.5.0
paramiko==2.9.2
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.7.0
pluggy==1.0.0
poyo==0.5.0
prettytable==3.2.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyface==7.4.1
pyflakes==2.4.0
pygame==2.1.2
Pygments==2.11.2
pylint==2.12.2
pyls-spyder==0.4.0
pymongo==4.1.0
PyNaCl==1.5.0
pynndescent==0.5.6
pyparsing==3.0.7
PyQt5==5.12.3
PyQt5-sip==12.9.1
PyQtWebEngine==5.12.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-lsp-black==1.1.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-slugify==6.1.1
PythonWebHDFS==0.2.3
pytz==2021.3
PyWavelets==1.3.0
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.1.1
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.20.0
rope==0.23.0
rsa==4.8
Rtree==0.9.7
scanpy==1.8.2
schema==0.7.5
SciencePlots==1.0.9
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
simplejson==3.17.6
sinfo==0.3.4
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.2.2
spyder-kernels==2.2.1
spyder-memory-profiler==0.2.1
statsmodels==0.13.2
stdlib-list==0.8.0
sweetviz==2.1.3
tables==3.6.1
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
textdistance==4.2.2
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
three-merge==0.1.1
tifffile==2021.11.2
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.10.0
toolz==0.11.2
torch==1.8.1
torch-geometric==2.0.4
torchfile==0.1.0
torchvision==0.9.1
torchxrayvision==0.0.37
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
traits==6.3.2
traitsui==7.3.1
typed-ast==1.5.2
typeguard==2.13.3
typing_extensions==4.1.1
ujson==5.1.0
umap-learn==0.5.2
urllib3==1.26.9
vim==0.0.1
visdom==0.1.8.9
vtk==9.1.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
websockets==10.2
Werkzeug==2.1.1
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.14.0
wslink==1.4.3
xgboost==1.5.2
xlrd==1.2.0
xshell==1.4.1
yapf==0.32.0
yarl==1.7.2
zipp==3.8.0
